+++ haul/__init__.py
@@ -1,6 +1,6 @@
 # coding: utf-8
 
-__version__ = '1.3.0'
+__version__ = '1.3.1'
 
 from .api import find_images
 from .core import Haul, HaulResult